Linting Initialization

+++ tests/test_metastore.py
@@ -60,8 +60,10 @@ class TestMetastore(TestCase):
     assert path.exists(self.td.test_path)
     assert DeltaTable.is_deltatable(self.td.test_path)
 
-    # Post-assertion cleanup
-    rmtree(self.td.test_path)
+    # Clear table if it exists
+    if path.exists(self.td.test_path):
+      rmtree(self.td.test_path)
+    assert not path.exists(self.td.test_path)
 
   def test_list_domains(self) -> None:
     # Assert metastore can return the list of domains
